+++ crates/katana/cli/src/lib.rs
@@ -1,8 +1,10 @@
 #![cfg_attr(not(test), warn(unused_crate_dependencies))]
 
+pub mod file;
 pub mod node;
 pub mod options;
 pub mod utils;
 
-pub use node::{NodeArgs, NodeArgsConfig};
+pub use file::NodeArgsConfig;
+pub use node::NodeArgs;
 pub use options::*;
